Move pmap to bin/glusterd2/pmap
